Pin the CM_3037 Tag forward to int
`cond ? DefMsg.Tag : 0` mixes ushort with the int literal 0, which the conditional
operator can resolve in either direction (ushort widens to int; the constant 0
narrows to ushort). Spell out the int so the SendMsg/SendActionMsg overload cannot
be reached through a narrowing of the Tag.

+++ GameSvr/UsrSystem/UsrEngn.cs
@@ -2620,13 +2620,13 @@ namespace GameSvr
                     {
                         PlayObject.SendActionMsg(PlayObject, DefMsg.Ident, DefMsg.Series & 7,
                             DefMsg.Recog, DefMsg.Param,
-                            DefMsg.Ident == Grobal2.CM_3037 ? DefMsg.Tag : 0, "");
+                            DefMsg.Ident == Grobal2.CM_3037 ? (int)DefMsg.Tag : 0, "");
                     }
                     else
                     {
                         PlayObject.SendMsg(PlayObject, DefMsg.Ident, DefMsg.Series & 7,
                             DefMsg.Recog, DefMsg.Param,
-                            DefMsg.Ident == Grobal2.CM_3037 ? DefMsg.Tag : 0, "");
+                            DefMsg.Ident == Grobal2.CM_3037 ? (int)DefMsg.Tag : 0, "");
                     }
                     break;
                 case Grobal2.CM_SAY:
